# docs/openapi.yaml
components:
    schemas:
...
        dto.CreateTransactionRequest:
            properties:
                amount:
                    type: integer
                category_id:
                    type: string
                date:
                    type: string
                name:
                    type: string
                payment_method_id:
                    type: string
            required:
                - amount
                - date
                - name
                - payment_method_id
            type: object
...
        dto.ErrorItem:
            properties:
                name:
                    type: string
                reason:
                    type: string
            type: object
        dto.ErrorResponse:
            properties:
                errors:
                    items:
                        $ref: '#/components/schemas/dto.ErrorItem'
                    type: array
                message:
                    type: string
            type: object
...
        dto.GetTransactionResponse:
            properties:
                account_id:
                    type: string
                amount:
                    type: integer
                category_id:
                    type: string
                category_name:
                    type: string
                created_at:
                    type: string
                date:
                    type: string
                deleted_at:
                    type: string
                external_id:
                    type: string
                id:
                    type: string
                institution_id:
                    type: string
                institution_logo:
                    type: string
                institution_name:
                    type: string
                is_ignored:
                    type: boolean
                name:
                    type: string
                payment_method_id:
                    type: string
                payment_method_name:
                    type: string
                search_document: {}
                updated_at:
                    type: string
                user_id:
                    type: string
            type: object
...
        dto.ListTransactionCategoriesResponse:
            properties:
                items:
                    items:
                        $ref: '#/components/schemas/entity.TransactionCategory'
                    type: array
                page:
                    type: integer
                page_size:
                    type: integer
                total_items:
                    type: integer
                total_pages:
                    type: integer
            type: object
        dto.ListTransactionsResponse:
            properties:
                items:
                    items:
                        $ref: '#/components/schemas/entity.FullTransaction'
                    type: array
                page:
                    type: integer
                page_size:
                    type: integer
                total_items:
                    type: integer
                total_pages:
                    type: integer
            type: object
...
        dto.UpdateTransactionRequest:
            properties:
                account_id:
                    type: string
                amount:
                    type: integer
                category_id:
                    type: string
                date:
                    type: string
                id:
                    type: string
                institution_id:
                    type: string
                name:
                    type: string
                payment_method_id:
                    type: string
            required:
                - id
            type: object
...
        entity.FullTransaction:
            properties:
                account_id:
                    type: string
                amount:
                    type: integer
                category_id:
                    type: string
                category_name:
                    type: string
                created_at:
                    type: string
                date:
                    type: string
                deleted_at:
                    type: string
                external_id:
                    type: string
                id:
                    type: string
                institution_id:
                    type: string
                institution_logo:
                    type: string
                institution_name:
                    type: string
                is_ignored:
                    type: boolean
                name:
                    type: string
                payment_method_id:
                    type: string
                payment_method_name:
                    type: string
                search_document: {}
                updated_at:
                    type: string
                user_id:
                    type: string
            type: object
...
        entity.TransactionCategory:
            properties:
                created_at:
                    type: string
                deleted_at:
                    type: string
                external_id:
                    type: string
                id:
                    type: string
                name:
                    type: string
                updated_at:
                    type: string
            type: object
...
    securitySchemes:
        BasicAuth:
            scheme: basic
            type: http
        BearerAuth:
            description: Type "Bearer" followed by a space and JWT token.
            in: header
            name: Authorization
            type: apiKey
info:
    contact:
        email: danielmesquitta123@gmail.com
        name: Daniel Mesquita
    description: API Finance Manager
    title: API Finance Manager
    version: "1.0"
openapi: 3.0.3
paths:
...
    /v1/admin/transactions/categories/sync:
        post:
            description: Sync categories from open finance
            responses:
                "204":
                    description: No Content
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/dto.ErrorResponse'
                    description: Internal Server Error
            security:
                - BasicAuth: []
            summary: Sync categories from open finance
            tags:
                - Category
    /v1/admin/transactions/sync:
        post:
            description: Webhook to sync transactions from open finance
            parameters:
                - description: User IDs
                  in: query
                  name: user_ids
                  schema:
                    items:
                        type: string
                    type: array
            responses:
                "204":
                    description: No Content
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/dto.ErrorResponse'
                    description: Bad Request
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/dto.ErrorResponse'
                    description: Internal Server Error
            security:
                - BasicAuth: []
            summary: Sync transactions from open finance
            tags:
                - Transaction
...
    /v1/transactions:
        get:
            description: List transactions
            parameters:
                - description: Search
                  in: query
                  name: search
                  schema:
                    type: string
                - description: Page
                  in: query
                  name: page
                  schema:
                    type: integer
                - description: Page size
                  in: query
                  name: page_size
                  schema:
                    type: integer
                - description: Start date
                  in: query
                  name: start_date
                  schema:
                    format: date
                    type: string
                - description: End date
                  in: query
                  name: end_date
                  schema:
                    format: date
                    type: string
                - description: Institution IDs
                  in: query
                  name: institution_ids
                  schema:
                    items:
                        type: string
                    type: array
                - description: Category IDs
                  in: query
                  name: category_ids
                  schema:
                    items:
                        type: string
                    type: array
                - description: Payment method IDs
                  in: query
                  name: payment_method_ids
                  schema:
                    items:
                        type: string
                    type: array
                - description: Filter only expenses
                  in: query
                  name: is_expense
                  schema:
                    type: boolean
                - description: Filter only incomes
                  in: query
                  name: is_income
                  schema:
                    type: boolean
                - description: Filter ignored or not ignored transactions
                  in: query
                  name: is_ignored
                  schema:
                    type: boolean
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/dto.ListTransactionsResponse'
                    description: OK
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/dto.ErrorResponse'
                    description: Unauthorized
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/dto.ErrorResponse'
                    description: Internal Server Error
            security:
                - BearerAuth: []
            summary: List transactions
            tags:
                - Transaction
        post:
            description: Create transactions
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/dto.CreateTransactionRequest'
                description: Request body
                required: true
                x-originalParamName: request
            responses:
                "201":
                    description: Created
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/dto.ErrorResponse'
                    description: Bad Request
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/dto.ErrorResponse'
                    description: Internal Server Error
            security:
                - BearerAuth: []
            summary: Create transactions
            tags:
                - Transaction
    /v1/transactions/{transaction_id}:
        get:
            description: Get transaction
            parameters:
                - description: Transaction ID
                  in: path
                  name: transaction_id
                  required: true
                  schema:
                    format: uuid
                    type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/dto.GetTransactionResponse'
                    description: OK
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/dto.ErrorResponse'
                    description: Unauthorized
                "404":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/dto.ErrorResponse'
                    description: Not Found
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/dto.ErrorResponse'
                    description: Internal Server Error
            security:
                - BearerAuth: []
            summary: Get transaction
            tags:
                - Transaction
        put:
            description: Update transaction
            parameters:
                - description: Transaction ID
                  in: path
                  name: transaction_id
                  required: true
                  schema:
                    format: uuid
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/dto.UpdateTransactionRequest'
                description: Request body
                required: true
                x-originalParamName: request
            responses:
                "204":
                    description: No Content
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/dto.ErrorResponse'
                    description: Unauthorized
                "404":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/dto.ErrorResponse'
                    description: Not Found
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/dto.ErrorResponse'
                    description: Internal Server Error
            security:
                - BearerAuth: []
            summary: Update transaction
            tags:
                - Transaction
    /v1/transactions/categories:
        get:
            description: List categories
            parameters:
                - description: Search
                  in: query
                  name: search
                  schema:
                    type: string
                - description: Page
                  in: query
                  name: page
                  schema:
                    type: integer
                - description: Page size
                  in: query
                  name: page_size
                  schema:
                    type: integer
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/dto.ListTransactionCategoriesResponse'
                    description: OK
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/dto.ErrorResponse'
                    description: Unauthorized
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/dto.ErrorResponse'
                    description: Internal Server Error
            security:
                - BearerAuth: []
            summary: List categories
            tags:
                - Category
...